requirements.txt
cleverhans==3.1.0
matplotlib==3.4.2
numpy==1.19.5
pandas==1.2.4
Pillow==8.2.0
scikit-learn==0.24.2
scipy==1.6.3
tensorflow===2.5.0
torch==1.8.1
torchvision===0.9.1
umap-learn==0.5.1
